+++ lib/yt/models/match_policy.rb
@@ -13,7 +13,7 @@ module Yt
 
       def update(attributes = {})
         underscore_keys! attributes
-        do_patch body: {policyId: attributes[:policy_id]}
+        do_patch body: attributes.slice(:policy_id)
         true
       end
 
@@ -25,7 +25,7 @@ module Yt
       def patch_params
         super.tap do |params|
           params[:path] = "/youtube/partner/v1/assets/#{@asset_id}/matchPolicy"
-          params[:params] = {onBehalfOfContentOwner: @auth.owner_name}
+          params[:params] = {on_behalf_of_content_owner: @auth.owner_name}
           params[:expected_response] = Net::HTTPOK
         end
       end