+++ lib/yt/models/request.rb
@@ -15,7 +15,6 @@ module Yt
     class Request
       def initialize(options = {})
         @auth = options[:auth]
-        @body = options[:body]
         @file = options[:file]
         @body_type = options.fetch :body_type, :json
         @expected_response = options.fetch :expected_response, Net::HTTPSuccess
@@ -24,8 +23,13 @@ module Yt
         @host = options.fetch :host, google_api_host
         @method = options.fetch :method, :get
         @path = options[:path]
-        @query = options.fetch(:params, {}).to_param
+        @body = options[:body]
+        camelize_keys! @body if options.fetch(:camelize_body, true)
+        params = options.fetch :params, {}
+        camelize_keys! params if options.fetch(:camelize_params, true)
+        @query = params.to_param
       end
+
 
       def run
         p as_curl if Yt.configuration.developing?
@@ -130,6 +134,13 @@ module Yt
           when :xml then Hash.from_xml body
           when :json then JSON body
         end if body
+      end
+
+      def camelize_keys!(object)
+        return object unless object.is_a?(Hash)
+        object.dup.each_key do |key|
+          object[key.to_s.camelize(:lower).to_sym] = object.delete key
+        end
       end
 
       # There are two cases to run a request again: YouTube responds with a